refactor: 파일 업로드 로직을 훅으로 분리
